Refactor the augment task, remove it from it from the graph database & from post-process.
