+++ src/AAB.EBA/Blockchains/Bitcoin/OffChain/UtxoLifecycleEconomics.cs
@@ -3,9 +3,9 @@ using AAB.EBA.Graph.Db.Neo4jDb;
 using AAB.EBA.Utilities;
 using System.IO.Compression;
 
-namespace AAB.EBA.Blockchains.Bitcoin.Utilities;
+namespace AAB.EBA.Blockchains.Bitcoin.OffChain;
 
-public class UtxoEconomics(ILogger<BitcoinOrchestrator> logger)
+public class EconomicAugmentor(ILogger<BitcoinOrchestrator> logger, Options options)
 {
     private class BatchUtxoDeltas(int maximumBlockHeight)
     {
@@ -40,14 +40,32 @@ public class UtxoEconomics(ILogger<BitcoinOrchestrator> logger)
         }
     }
 
+    private readonly Options _options = options;
     private readonly ILogger<BitcoinOrchestrator> _logger = logger;
 
     private readonly int _creationHeightIdx = T2SEdgeDescriptor.StaticMapper.GetPropertyCsvIndex(x => x.CreationHeight);
     private readonly int _spentHeightIdx = T2SEdgeDescriptor.StaticMapper.GetPropertyCsvIndex(x => x.SpentHeight);
     private readonly int _valueIdx = T2SEdgeDescriptor.StaticMapper.GetPropertyCsvIndex(x => x.Value);
 
-    public async Task SetBlockOnChainEconomicsAsync(
-        List<Batch> batches,
+    public async Task SetBlockMarketIndicatorsAsync(CancellationToken ct)
+    {
+        var batches = await Batch.DeserializeBatchesAsync(_options.Bitcoin.MapSpends.BatchesFilename);
+        _logger.LogInformation(
+            "Deserialized {n:n0} batches from {filename}.", 
+            batches.Count, _options.Bitcoin.MapSpends.BatchesFilename);
+
+        var (_, blockNodes) = await Utilities.BitcoinHelpers.GetHeightToBatchMapping(batches, _logger, ct);
+
+        var blockOhlcvMapping = await SetBlockOHLCV(blockNodes, ct);
+
+        await SetBlockValuationMetrics(batches, blockNodes, blockOhlcvMapping, ct);
+
+        ComputeThermocap(blockNodes, blockOhlcvMapping);
+
+        await SaveBlockNodesWithUpdatedEconomicIndicators(batches, blockNodes, ct);
+    }
+
+    private async Task SetBlockValuationMetrics(List<Batch> batches,
         SortedDictionary<long, BlockNode> blockNodes,
         Dictionary<long, OHLCV> ohlcv,
         CancellationToken ct)
@@ -288,5 +306,133 @@ public class UtxoEconomics(ILogger<BitcoinOrchestrator> logger)
 
         blockNode.BlockMetadata.UnrealizedLoss = totalUnrealizedLoss > 0 ? (decimal)totalUnrealizedLoss : 0m;
         blockNode.BlockMetadata.UnrealizedProfit = totalUnrealizedProfit > 0 ? (decimal)totalUnrealizedProfit : 0m;
+    }
+
+    private void ComputeThermocap(
+        SortedDictionary<long, BlockNode> blockNodes,
+        Dictionary<long, OHLCV> blockOHLCVMapping)
+    {
+        _logger.LogInformation("Computing thermocap for blocks.");
+
+        foreach (var block in blockNodes)
+        {
+            if (block.Key == 0)
+                continue;
+
+            if (blockOHLCVMapping.TryGetValue(block.Value.BlockMetadata.Height, out var ohlcv))
+            {
+                var prevThermocap = blockNodes[block.Key - 1].BlockMetadata.Thermocap ?? 0;
+                block.Value.BlockMetadata.Thermocap =
+                    prevThermocap +
+                    ohlcv.GetFiatValue(block.Value.TripletTypeValueSum[C2TEdge.Kind]);
+            }
+        }
+
+        _logger.LogInformation("Completed computing thermocap for blocks.");
+    }
+
+
+    private async Task<Dictionary<long, OHLCV>> SetBlockOHLCV(SortedDictionary<long, BlockNode> blockNodes, CancellationToken ct)
+    {
+        if (!OHLCV.TryParseFile(_options.Bitcoin.Augmentor.BlockOhlcvMappedFilename, out var blockOhlcvMapping))
+        {
+            _logger.LogError(
+                "Failed to parse OHLCV data from file: {filename}",
+                _options.Bitcoin.Augmentor.BlockOhlcvMappedFilename);
+            return [];
+        }
+        else
+        {
+            _logger.LogInformation(
+                "Successfully parsed OHLCV data for {count:n0} blocks from file: {filename}",
+                blockOhlcvMapping.Count,
+                _options.Bitcoin.Augmentor.BlockOhlcvMappedFilename);
+        }
+
+        ct.ThrowIfCancellationRequested();
+
+        var counter = 0;
+        foreach (var block in blockNodes)
+        {
+            if (blockOhlcvMapping.TryGetValue(block.Key, out var ohlcv))
+            {
+                counter++;
+                block.Value.BlockMetadata.Ohlcv = ohlcv;
+            }
+        }
+
+        _logger.LogInformation(
+            "Extended {n:n0}/{total:n0} block nodes with OHLCV data; did not find mapping OHLCV for {d:n0} block nodes.",
+            counter, blockNodes.Count, blockNodes.Count - counter);
+
+        return blockOhlcvMapping;
+    }
+
+    private async Task SaveBlockNodesWithUpdatedEconomicIndicators(List<Batch> batches, SortedDictionary<long, BlockNode> blockNodes, CancellationToken ct)
+    {
+        var prefix = BlockNodeDescriptor.OhlcvPropNamePrefix;
+        var mapper = BlockNodeDescriptor.StaticMapper;
+        var heightIdx = mapper.GetPropertyCsvIndex(x => x.BlockMetadata.Height);
+        var oOpenIdx = mapper.GetPropertyCsvIndex(x => x.BlockMetadata.Ohlcv.Open, propertyNamePrefix: prefix);
+        var oCloseIdx = mapper.GetPropertyCsvIndex(x => x.BlockMetadata.Ohlcv.Close, propertyNamePrefix: prefix);
+        var oHighIdx = mapper.GetPropertyCsvIndex(x => x.BlockMetadata.Ohlcv.High, propertyNamePrefix: prefix);
+        var oLowIdx = mapper.GetPropertyCsvIndex(x => x.BlockMetadata.Ohlcv.Low, propertyNamePrefix: prefix);
+        var oHlc4Idx = mapper.GetPropertyCsvIndex(x => x.BlockMetadata.Ohlcv.OHLC4, propertyNamePrefix: prefix);
+        var oVolumeIdx = mapper.GetPropertyCsvIndex(x => x.BlockMetadata.Ohlcv.Volume, propertyNamePrefix: prefix);
+        var oVwapIdx = mapper.GetPropertyCsvIndex(x => x.BlockMetadata.Ohlcv.VWAP, propertyNamePrefix: prefix);
+
+        var realizedCapIdx = mapper.GetPropertyCsvIndex(x => x.BlockMetadata.RealizedCap);
+        var unrealizedProfitIdx = mapper.GetPropertyCsvIndex(x => x.BlockMetadata.UnrealizedProfit);
+        var unrealizedLossIdx = mapper.GetPropertyCsvIndex(x => x.BlockMetadata.UnrealizedLoss);
+
+        var marketCapIdx = mapper.GetPropertyCsvIndex(x => x.BlockMetadata.MarketCap);
+        var nuplIdx = mapper.GetPropertyCsvIndex(x => x.BlockMetadata.NUPL);
+        var nulIdx = mapper.GetPropertyCsvIndex(x => x.BlockMetadata.NUL);
+        var nupIdx = mapper.GetPropertyCsvIndex(x => x.BlockMetadata.NUP);
+        var mvrvIdx = mapper.GetPropertyCsvIndex(x => x.BlockMetadata.MVRV);
+        var thermoIdx = mapper.GetPropertyCsvIndex(x => x.BlockMetadata.Thermocap);
+
+        await Parallel.ForEachAsync(batches, new ParallelOptions { MaxDegreeOfParallelism = 1, CancellationToken = ct, }, async (batch, ct) =>
+        {
+            using var reader = new StreamReader(new GZipStream(
+                File.OpenRead(batch.GetFilename(BlockNode.Kind)),
+                CompressionMode.Decompress));
+
+            using var writer = new StreamWriter(new GZipStream(
+                File.Create(Helpers.AddPostfixToFilename(batch.GetFilename(BlockNode.Kind), "_with_economic_indicators")),
+                CompressionLevel.Optimal));
+
+            string? line;
+
+            while ((line = await reader.ReadLineAsync(ct)) != null)
+            {
+                var cols = line.Split(Options.CsvDelimiter);
+                var height = long.Parse(cols[heightIdx]);
+                var blockNode = blockNodes[height];
+                if (blockNode.BlockMetadata.Ohlcv != null)
+                {
+                    cols[oOpenIdx] = blockNode.BlockMetadata.Ohlcv.Open.ToString();
+                    cols[oCloseIdx] = blockNode.BlockMetadata.Ohlcv.Close.ToString();
+                    cols[oHighIdx] = blockNode.BlockMetadata.Ohlcv.High.ToString();
+                    cols[oLowIdx] = blockNode.BlockMetadata.Ohlcv.Low.ToString();
+                    cols[oHlc4Idx] = blockNode.BlockMetadata.Ohlcv.OHLC4.ToString();
+                    cols[oVolumeIdx] = blockNode.BlockMetadata.Ohlcv.Volume.ToString();
+                    cols[oVwapIdx] = blockNode.BlockMetadata.Ohlcv.VWAP.ToString();
+
+                    cols[realizedCapIdx] = blockNode.BlockMetadata.RealizedCap?.ToString() ?? "0";
+                    cols[unrealizedProfitIdx] = blockNode.BlockMetadata.UnrealizedProfit?.ToString() ?? "0";
+                    cols[unrealizedLossIdx] = blockNode.BlockMetadata.UnrealizedLoss?.ToString() ?? "0";
+
+                    cols[marketCapIdx] = blockNode.BlockMetadata.MarketCap?.ToString() ?? "0";
+                    cols[nuplIdx] = blockNode.BlockMetadata.NUPL?.ToString() ?? "0";
+                    cols[nulIdx] = blockNode.BlockMetadata.NUL?.ToString() ?? "0";
+                    cols[nupIdx] = blockNode.BlockMetadata.NUP?.ToString() ?? "0";
+                    cols[mvrvIdx] = blockNode.BlockMetadata.MVRV?.ToString() ?? "0";
+                    cols[thermoIdx] = blockNode.BlockMetadata.Thermocap?.ToString() ?? "0";
+                }
+
+                await writer.WriteLineAsync(string.Join(Options.CsvDelimiter, cols));
+            }
+        });
     }
 }
