+++ src/AAB.EBA/CLI/CLI.cs
@@ -238,11 +238,6 @@ internal class Cli
                 "(reference: https://neo4j.com/blog/bulk-data-import-neo4j-3-0/)"
         };
 
-        var blockMarketMappingOption = new Option<string>("--map-block-market")
-        {
-            Required = false
-        };
-
         var cmd = new Command(
             name: "traverse",
             description: "Traverses the blockchain in the defined range collecting the set metrics.")
@@ -256,8 +251,7 @@ internal class Cli
             txoFilenameOption,
             skipGraphSerialization,
             trackTxoOption,
-            maxEntriesPerBatch,
-            blockMarketMappingOption
+            maxEntriesPerBatch
         };
 
 
@@ -307,8 +301,7 @@ internal class Cli
                 trackTxoOption: trackTxoOption,
                 txoFilenameOption: txoFilenameOption,
                 skipGraphSerializationOption: skipGraphSerialization,
-                maxEntriesPerBatch: maxEntriesPerBatch,
-                blockMarketMappingOption: blockMarketMappingOption);
+                maxEntriesPerBatch: maxEntriesPerBatch);
 
             try
             {
@@ -674,14 +667,23 @@ internal class Cli
 
     private Command GetBitcoinAugmentCmd(Options defaultOptions, Func<Options, Task> handlerAsync)
     {
+        var batchesFilenameOption = new Option<string>("--batches-filename")
+        {
+            Required = true
+        };
+
         var ohlcvFilenameOption = new Option<string>("--ohlcv-filename")
         {
             Description = "A TSV file containing block metadata (height, median time) mapped to aggregated OHLCV market data."
         };
         var cmd = new Command(
             name: "augment",
-            description: "Augments the graph with additional features, such as NUPL, using the output of the map-market command and other similar files.")
+            description: 
+            "Computes UTxO-based economic metrics (e.g., realized cap, unrealized profit/loss, and NUPL) " +
+            "to augment the graph data. You may run this optional command after the 'post-traverse' command " +
+            "and before importing the graph into a graph database.")
         {
+            batchesFilenameOption,
             ohlcvFilenameOption
         };
         cmd.SetAction(async (parseResult, cancellationToken) =>
@@ -690,7 +692,8 @@ internal class Cli
                 parseResult,
                 workingDirOption: _workingDirOption,
                 statusFilenameOption: _statusFilenameOption,
-                augmentroOhlcvOption: ohlcvFilenameOption);
+                augmentroOhlcvOption: ohlcvFilenameOption,
+                batchesFilenameOption: batchesFilenameOption);
             try
             {
                 await handlerAsync(options);
@@ -710,12 +713,6 @@ internal class Cli
             Required = true
         };
 
-        var ohlcvFilenameOption = new Option<string>("--ohlcv-filename")
-        {
-            Required = false,
-            Description = "A TSV file containing block metadata (height, median time) mapped to aggregated OHLCV market data."
-        };
-
         var cmd = new Command(
             name: "post-traverse",
             description:
@@ -723,8 +720,7 @@ internal class Cli
                 "Computes node and edge properties that depend on cross-block information, " +
                 "such as UTxO spending and total Bitcoin supply per block.")
         {
-            batchesFilenameOption,
-            ohlcvFilenameOption
+            batchesFilenameOption
         };
 
         cmd.SetAction(async (parseResult, cancellationToken) =>
@@ -733,8 +729,7 @@ internal class Cli
                 parseResult,
                 workingDirOption: _workingDirOption,
                 statusFilenameOption: _statusFilenameOption,
-                batchesFilenameOption: batchesFilenameOption,
-                augmentroOhlcvOption: ohlcvFilenameOption);
+                batchesFilenameOption: batchesFilenameOption);
 
             try
             {
