+++ src/AAB.EBA/Graph/Db/Neo4jDb/Neo4jDb.cs
@@ -310,96 +310,6 @@ public class Neo4jDb<T> : IGraphDb<T> where T : GraphBase
         return edgeCount;
     }
 
-    // TODO: this method should not be here;
-    // it is specific to the Bitcoin graph model
-    // and this class should be kept agnostic to the graph model.
-    public async Task SetRealizedCap(
-        SortedDictionary<long, BlockNode> blockNodes,
-        Dictionary<long, OHLCV> ohlcv,
-        CancellationToken ct)
-    {
-        if (blockNodes.Count == 0)
-            return;
-
-        var sortedHeights = blockNodes.Keys.ToArray();
-        var minHeight = sortedHeights[0];
-        var maxHeight = sortedHeights[^1];
-
-        foreach (var block in blockNodes.Values)
-            block.BlockMetadata.RealizedCap = 0;
-
-        await VerifyConnectivityAsync(ct);
-        await using var session = _driver.AsyncSession(x => x.WithDefaultAccessMode(AccessMode.Read));
-
-        var edgeCounter = 0;
-        double totalEdgeCount = await GetEdgeTypeCount(T2SEdge.Kind.Relation, ct);
-        var processedEdgeCount = 0;
-        var skippedEdgeCounter = 0;
-
-        _logger.LogInformation("Starting iteration through {b} edges.", T2SEdge.Kind.Relation);
-
-        await session.ExecuteReadAsync(async tx =>
-        {
-            var cursor = await tx.RunAsync(
-                $"MATCH (tx)-[r:{T2SEdge.Kind.Relation}]->(script) " +
-                $"WHERE " +
-                $"r.{nameof(T2SEdge.CreationHeight)} <= {maxHeight} " +
-                $"AND " +
-                $"r.{nameof(T2SEdge.SpentHeight)} > {minHeight} " +
-                $"AND " +
-                $"script.{nameof(ScriptNode.ScriptType)} <> '{ScriptType.NullData}' " +
-                $"RETURN " +
-                $"r.{nameof(T2SEdge.CreationHeight)} AS creationHeight, " +
-                $"r.{nameof(T2SEdge.SpentHeight)} AS spentHeight, " +
-                $"r.{nameof(T2SEdge.Value)} AS value");
-
-            while (await cursor.FetchAsync())
-            {
-                ct.ThrowIfCancellationRequested();
-                edgeCounter++;
-
-                var creationHeight = cursor.Current["creationHeight"].As<long>();
-                var spentHeight = cursor.Current["spentHeight"].As<long>();
-                var value = cursor.Current["value"].As<long>();
-
-                if (!ohlcv.TryGetValue(creationHeight, out var creationBlockOHLCV))
-                {
-                    skippedEdgeCounter++;
-                }
-                else
-                {
-                    var utxoFiatValueAtCreation = creationBlockOHLCV.GetFiatValue(value);
-                    foreach (var h in sortedHeights.GetViewBetween(creationHeight, spentHeight))
-                    {
-                        var block = blockNodes[h];
-                        block.BlockMetadata.RealizedCap += utxoFiatValueAtCreation;
-
-                        if (block.BlockMetadata.Ohlcv != null)
-                        {
-                            if (utxoFiatValueAtCreation < block.BlockMetadata.Ohlcv.VWAP)
-                                block.BlockMetadata.UnrealizedLoss += block.BlockMetadata.Ohlcv.VWAP - utxoFiatValueAtCreation;
-                            else
-                                block.BlockMetadata.UnrealizedProfit += utxoFiatValueAtCreation - block.BlockMetadata.Ohlcv.VWAP;
-                        }
-                    }
-
-                    processedEdgeCount++;
-                }
-
-                if (edgeCounter % 100000 == 0)
-                {
-                    _logger.LogInformation(
-                        "Traversed {e:n0} / {t:n0} edges. Processed {p:n0} edges and skipped {s:n0} edges due to missing related OHLCV data.",
-                        edgeCounter, totalEdgeCount, processedEdgeCount, skippedEdgeCounter);
-                }
-            }
-        });
-
-        _logger.LogInformation(
-            "Traversed {e:n0} / {t:n0} edges. Processed {p:n0} edges and skipped {s:n0} edges due to missing related OHLCV data.",
-            edgeCounter, totalEdgeCount, processedEdgeCount, skippedEdgeCounter);
-    }
-
     public async Task BulkUpdateNodePropertiesAsync(
         NodeKind label,
         string idProperty,
